% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize_col.r
\name{vectorize_col}
\alias{vectorize_col}
\title{A surveyCleanR function}
\usage{
vectorize_col(df, col_to_vect, colname)
}
\arguments{
\item{df}{Required Dataframe object.}

\item{col_to_vect}{Required column/vector object.}

\item{colname}{Required column name}
}
\description{
This function allows you to deal with missing data in a column and quantify variables.
}
\examples{
vectorize_col(dataframe,column_in_frame,name_of_column)
}
\keyword{column}
\keyword{dataframe}
\keyword{survey}
