% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize_df.r
\name{vectorize_df}
\alias{vectorize_df}
\title{A surveyCleanR function}
\usage{
vectorize_df(df)
}
\arguments{
\item{df}{Required Dataframe object.}
}
\description{
This function allows you to deal with missing data in a column and quantify variables.
}
\examples{
vectorize_df(df_survey)
}
\keyword{dataframe}
\keyword{survey}
